Fix: Correct clinical CNA reporting from fold changes to categorical values
This fixes an issue where categorical CNAs for clinical samples were incorrectly reported as fold changes.
The get_cna_calls.py script now properly processes CNA categories represented as numbers,
converting them to 'AMP' if the value is 2 or 'DeepDel' if the value is -1.5 or -2.
A new column 'cna_type' has been added to the final CNA file for this